# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "The Garden Of Rememberance",
        "description": "",
        "version": "3.0.0"
    },
    "paths": {
...
        "/api/report": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get list Report",
                "operationId": "App\\Http\\Controllers\\ReportController::listReport",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit page",
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter by title",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Error server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Create Report",
                "operationId": "App\\Http\\Controllers\\ReportController::create",
                "parameters": [
                    {
                        "name": "start_time",
                        "in": "query",
                        "description": "Start Time",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "description": "End Time",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "service",
                        "in": "query",
                        "description": "Service",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Error server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reports"
                ],
                "summary": "Delete Reports file",
                "operationId": "deleteReportsFile",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "ids"
                                ],
                                "properties": {
                                    "ids": {
                                        "description": "ID Report File",
                                        "type": "string",
                                        "default": "[1,2,3]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Error server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/generate-report/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Generate Report",
                "operationId": "App\\Http\\Controllers\\ReportController::generateCollections",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Report",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Error server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}